tweaks & mouse recentering on decapture

+++ Outrospection/Types.h
@@ -59,7 +59,7 @@ struct FontCharacter
 	GLuint textureId;
 	glm::ivec2 size;
 	glm::ivec2 bearing; // offset from base line
-	unsigned int advance; // offset to advance to next glyph
+	long advance; // offset to advance to next glyph
 };
 
 struct DummyObj
